moved InvoiceType in new package .specifications
