+++ src/module/src/main/java/com/model/invoice/Invoice.java
@@ -1,6 +1,7 @@
 package com.model.invoice;
 
 import com.model.customer.Customer;
+import com.model.invoice.specifications.InvoiceType;
 import com.model.product.AbstractProduct;
 
 import java.time.LocalDateTime;
