+++ src/module/src/main/java/com/model/invoice/specifications/InvoiceType.java
@@ -1,4 +1,4 @@
-package com.model.invoice;
+package com.model.invoice.specifications;
 
 public enum InvoiceType {
     wholesale, retail, low_age
